[UPDATE]post request

+++ domain/src/main/java/com/findapple/findapple/domain/object/LocationInfo.kt
@@ -1,3 +1,3 @@
 package com.findapple.findapple.domain.`object`
 
-data class LocationInfo(val longitude: Long, val latitude: Long)
+data class LocationInfo(val longitude: Double, val latitude: Double)
